requirements.txt
huggingface-hub
transformers
accelerate
sentinelhub
Pillow
rasterio
scipy